# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/geofences": {
      "get": {
        "operationId": "Operations_geofences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Geofences"
        ],
        "summary": "Retrieve geofences",
        "description": "Retrieve operation in the Geofences module. Organization scope is enforced from the authenticated session."
      },
      "post": {
        "operationId": "Operations_createGeofence",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeofenceDto"
              },
              "example": {
                "name": "مستودع الرياض",
                "type": "circle",
                "geometry": {
                  "center": {
                    "latitude": 24.7136,
                    "longitude": 46.6753
                  },
                  "radiusMeters": 1000
                },
                "color": "#28e1da",
                "active": true
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "name": "مستودع الرياض",
                    "type": "circle",
                    "geometry": {
                      "center": {
                        "latitude": 24.7136,
                        "longitude": 46.6753
                      },
                      "radiusMeters": 1000
                    },
                    "color": "#28e1da",
                    "active": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Geofences"
        ],
        "summary": "Create / execute geofences",
        "description": "Create / execute operation in the Geofences module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/geofences/{id}": {
      "patch": {
        "operationId": "Operations_updateGeofence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeofenceDto"
              },
              "example": {
                "name": "مستودع الرياض",
                "type": "circle",
                "geometry": {
                  "center": {
                    "latitude": 24.7136,
                    "longitude": 46.6753
                  },
                  "radiusMeters": 1500
                },
                "color": "#28e1da",
                "active": true
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "name": "مستودع الرياض",
                    "type": "circle",
                    "geometry": {
                      "center": {
                        "latitude": 24.7136,
                        "longitude": 46.6753
                      },
                      "radiusMeters": 1500
                    },
                    "color": "#28e1da",
                    "active": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Geofences"
        ],
        "summary": "Update geofences",
        "description": "Update operation in the Geofences module. Organization scope is enforced from the authenticated session."
      },
      "delete": {
        "operationId": "Operations_removeGeofence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Geofences"
        ],
        "summary": "Delete geofences",
        "description": "Delete operation in the Geofences module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/alerts": {
      "get": {
        "operationId": "Operations_alerts",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "OPEN"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Alerts"
        ],
        "summary": "Retrieve alerts",
        "description": "Retrieve operation in the Alerts module. Organization scope is enforced from the authenticated session."
      },
      "post": {
        "operationId": "Operations_createAlert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertDto"
              },
              "example": {
                "vehicleId": "vehicle-id",
                "type": "OVERSPEED",
                "title": "تجاوز السرعة",
                "message": "تجاوزت المركبة السرعة المسموحة",
                "severity": "WARNING",
                "metadata": {
                  "limit": 120,
                  "actual": 132
                }
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "vehicleId": "vehicle-id",
                    "type": "OVERSPEED",
                    "title": "تجاوز السرعة",
                    "message": "تجاوزت المركبة السرعة المسموحة",
                    "severity": "WARNING",
                    "metadata": {
                      "limit": 120,
                      "actual": 132
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Alerts"
        ],
        "summary": "Create / execute alerts",
        "description": "Create / execute operation in the Alerts module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/alerts/{id}/status": {
      "patch": {
        "operationId": "Operations_alertStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertStatusDto"
              },
              "example": {
                "status": "ACKNOWLEDGED"
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "status": "ACKNOWLEDGED"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Alerts"
        ],
        "summary": "Update alerts",
        "description": "Update operation in the Alerts module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/maintenance": {
      "get": {
        "operationId": "Operations_maintenance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Maintenance"
        ],
        "summary": "Retrieve maintenance",
        "description": "Retrieve operation in the Maintenance module. Organization scope is enforced from the authenticated session."
      },
      "post": {
        "operationId": "Operations_createMaintenance",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceDto"
              },
              "example": {
                "vehicleId": "vehicle-id",
                "name": "تغيير الزيت",
                "description": "زيت وفلتر",
                "status": "UPCOMING",
                "intervalKm": 10000,
                "dueAt": "2026-09-01T00:00:00.000Z",
                "dueOdometer": 50000,
                "cost": 450
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "vehicleId": "vehicle-id",
                    "name": "تغيير الزيت",
                    "description": "زيت وفلتر",
                    "status": "UPCOMING",
                    "intervalKm": 10000,
                    "dueAt": "2026-09-01T00:00:00.000Z",
                    "dueOdometer": 50000,
                    "cost": 450
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Maintenance"
        ],
        "summary": "Create / execute maintenance",
        "description": "Create / execute operation in the Maintenance module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/maintenance/{id}": {
      "patch": {
        "operationId": "Operations_updateMaintenance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceDto"
              },
              "example": {
                "vehicleId": "vehicle-id",
                "name": "تغيير الزيت",
                "status": "COMPLETED",
                "cost": 475
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "vehicleId": "vehicle-id",
                    "name": "تغيير الزيت",
                    "status": "COMPLETED",
                    "cost": 475
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Maintenance"
        ],
        "summary": "Update maintenance",
        "description": "Update operation in the Maintenance module. Organization scope is enforced from the authenticated session."
      },
      "delete": {
        "operationId": "Operations_removeMaintenance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Maintenance"
        ],
        "summary": "Delete maintenance",
        "description": "Delete operation in the Maintenance module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/fuel": {
      "get": {
        "operationId": "Operations_fuel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Fuel"
        ],
        "summary": "Retrieve fuel",
        "description": "Retrieve operation in the Fuel module. Organization scope is enforced from the authenticated session."
      },
      "post": {
        "operationId": "Operations_createFuel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FuelDto"
              },
              "example": {
                "vehicleId": "vehicle-id",
                "liters": 42.5,
                "pricePerLiter": 2.33,
                "totalCost": 99.03,
                "odometer": 45220,
                "station": "محطة الرياض",
                "filledAt": "2026-08-17T10:00:00.000Z"
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "vehicleId": "vehicle-id",
                    "liters": 42.5,
                    "pricePerLiter": 2.33,
                    "totalCost": 99.03,
                    "odometer": 45220,
                    "station": "محطة الرياض",
                    "filledAt": "2026-08-17T10:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Fuel"
        ],
        "summary": "Create / execute fuel",
        "description": "Create / execute operation in the Fuel module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/fuel/{id}": {
      "delete": {
        "operationId": "Operations_removeFuel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Fuel"
        ],
        "summary": "Delete fuel",
        "description": "Delete operation in the Fuel module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/tasks": {
      "get": {
        "operationId": "Operations_tasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "Retrieve tasks",
        "description": "Retrieve operation in the Tasks module. Organization scope is enforced from the authenticated session."
      },
      "post": {
        "operationId": "Operations_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDto"
              },
              "example": {
                "vehicleId": "vehicle-id",
                "title": "توصيل شحنة",
                "description": "من المستودع إلى الفرع",
                "status": "NEW",
                "priority": 2,
                "startsAt": "2026-08-18T08:00:00.000Z",
                "dueAt": "2026-08-18T12:00:00.000Z"
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "vehicleId": "vehicle-id",
                    "title": "توصيل شحنة",
                    "description": "من المستودع إلى الفرع",
                    "status": "NEW",
                    "priority": 2,
                    "startsAt": "2026-08-18T08:00:00.000Z",
                    "dueAt": "2026-08-18T12:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "Create / execute tasks",
        "description": "Create / execute operation in the Tasks module. Organization scope is enforced from the authenticated session."
      }
    },
...
  },
  "info": {
    "title": "ORBIT Fleet API",
    "description": "Complete fleet operations API: authentication, vehicles, TCP-normalized telemetry, drivers, geofences, alerts, maintenance, fuel, tasks, device commands, cameras, configuration, users, audit and reports.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access token returned by POST /v1/auth/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "HttpOnly browser session cookie"
      }
    },
    "schemas": {
...
      "GeofenceDto": {
        "type": "object",
        "required": [
          "name",
          "type",
          "geometry"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "مستودع الرياض"
          },
          "type": {
            "type": "string",
            "enum": [
              "circle",
              "polygon"
            ],
            "example": "circle"
          },
          "geometry": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "center": {
                "latitude": 24.7136,
                "longitude": 46.6753
              },
              "radiusMeters": 1000
            }
          },
          "color": {
            "type": "string",
            "default": "#28e1da"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "AlertStatusDto": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "ACKNOWLEDGED",
              "RESOLVED"
            ],
            "example": "ACKNOWLEDGED"
          }
        }
      },
...
    }
  }
}